+++ lib/src/effects/relativistic/relativistic_effect_configuration.dart
@@ -107,12 +107,57 @@ class SolidEdges {
 /// ```
 @immutable
 class RelativisticEffectConfiguration extends EffectConfiguration {
-  /// Creates a `RelativisticEffectConfiguration` instance with the specified properties.
+  /// Creates an instance of [RelativisticEffectConfiguration] with customizable parameters
+  /// for simulating particle physics with relativistic effects.
+  ///
+  /// This configuration adds properties like gravity, velocity, and material characteristics such as friction and restitution,
+  /// which impact how particles behave under simulated physical laws.
+  ///
+  /// - [gravity]: The gravitational force applied to the particles, influencing their motion.
+  /// - [particleConfiguration]: General particle configuration inherited from [EffectConfiguration], defining how particles are emitted and animated.
+  /// - [configurationOverrider]: A function to dynamically override default particle configurations during runtime.
+  /// - [maxDensity]: The maximum density of the particles, affecting their mass and interaction with other particles. Defaults to [Density.defaultDensity].
+  /// - [maxFriction]: The maximum friction applied to particles, slowing them down when in contact with surfaces. Defaults to [Friction.ice], representing minimal friction.
+  /// - [maxRestitution]: The maximum restitution, controlling how much energy is conserved in particle collisions (elasticity). Defaults to [Restitution.rubberBall], representing a bouncy collision.
+  /// - [maxVelocity]: The maximum velocity a particle can reach, which can be limited to simulate realistic speed constraints. Defaults to [Velocity.rainDrop].
+  /// - [minDensity]: The minimum density of the particles, determining their resistance to external forces. Defaults to [Density.defaultDensity].
+  /// - [minFriction]: The minimum friction applied to particles, which influences their ability to slide across surfaces. Defaults to [Friction.ice].
+  /// - [minRestitution]: The minimum restitution for particles, controlling how little energy is conserved in collisions. Defaults to [Restitution.rubberBall].
+  /// - [minVelocity]: The minimum velocity a particle can have. Defaults to [Velocity.rainDrop].
+  /// - [onlyInteractWithEdges]: If `true`, particles will only interact with the defined edges and not with each other. Defaults to `false`.
+  /// - [solidEdges]: Specifies solid boundaries for the particles' movement. By default, it uses [SolidEdges.all()], indicating that particles are constrained by edges on all sides.
+  ///
+  /// Inherited parameters from [EffectConfiguration]:
+  /// - [emitCurve]: Controls the timing of particle emission.
+  /// - [emitDuration]: The duration between particle emissions.
+  /// - [fadeInCurve]: Controls the fade-in animation for particles.
+  /// - [fadeOutCurve]: Controls the fade-out animation for particles.
+  /// - [foreground]: Whether the effect should be rendered in the foreground.
+  /// - [maxAngle]: The maximum angle (in degrees) for the particle trajectory.
+  /// - [maxBeginScale]: The maximum initial scale of particles.
+  /// - [maxEndScale]: The maximum final scale of particles.
+  /// - [maxFadeInThreshold]: The maximum opacity level at which particles will complete fading in.
+  /// - [maxFadeOutThreshold]: The maximum opacity level at which particles will start fading out.
+  /// - [maxOriginOffset]: The maximum offset for the particle emission origin.
+  /// - [maxParticleLifespan]: The maximum time a particle can exist before being removed.
+  /// - [minAngle]: The minimum angle (in degrees) for particle trajectory.
+  /// - [minBeginScale]: The minimum initial scale of particles.
+  /// - [minEndScale]: The minimum final scale of particles.
+  /// - [minFadeInThreshold]: The minimum opacity level at which particles will start fading in.
+  /// - [minFadeOutThreshold]: The minimum opacity level at which particles will start fading out.
+  /// - [minOriginOffset]: The minimum offset for the particle emission origin.
+  /// - [minParticleLifespan]: The minimum time a particle can exist before being removed.
+  /// - [origin]: The origin point for particle emission, relative to the top-left of the container.
+  /// - [particleCount]: The total number of particles that will be emitted over the lifetime of the effect.
+  /// - [particlesPerEmit]: The number of particles emitted at each emission event.
+  /// - [particleLayer]: The visual layer to which particles belong.
+  /// - [scaleCurve]: Controls the scaling of particles over time.
+  /// - [startDelay]: The delay before the particle effect starts after initiation.
+  /// - [trail]: Specifies any trail effect applied to particles as they move.
   const RelativisticEffectConfiguration({
     required this.gravity,
     required super.particleConfiguration,
     this.configurationOverrider,
-    this.hardEdges = const SolidEdges.all(),
     this.maxDensity = Density.defaultDensity,
     this.maxFriction = Friction.ice,
     this.maxRestitution = Restitution.rubberBall,
@@ -122,6 +167,7 @@ class RelativisticEffectConfiguration extends EffectConfiguration {
     this.minRestitution = Restitution.rubberBall,
     this.minVelocity = Velocity.rainDrop,
     this.onlyInteractWithEdges = false,
+    this.solidEdges = const SolidEdges.all(),
     super.emitCurve,
     super.emitDuration,
     super.fadeInCurve,
@@ -129,18 +175,18 @@ class RelativisticEffectConfiguration extends EffectConfiguration {
     super.foreground,
     super.maxAngle,
     super.maxBeginScale,
-    super.maxDuration,
+    super.maxParticleLifespan,
     super.maxEndScale,
     super.maxFadeInThreshold,
     super.maxFadeOutThreshold,
     super.maxOriginOffset,
     super.minAngle,
     super.minBeginScale,
-    super.minDuration,
     super.minEndScale,
     super.minFadeInThreshold,
     super.minFadeOutThreshold,
     super.minOriginOffset,
+    super.minParticleLifespan,
     super.origin,
     super.particleCount,
     super.particlesPerEmit,
@@ -157,9 +203,6 @@ class RelativisticEffectConfiguration extends EffectConfiguration {
 
   /// The gravitational force applied to the particles.
   final Gravity gravity;
-
-  /// Specifies which edges of the container are hard (solid) boundaries.
-  final SolidEdges hardEdges;
 
   /// The maximum density of the particles.
   final Density maxDensity;
@@ -188,13 +231,15 @@ class RelativisticEffectConfiguration extends EffectConfiguration {
   /// Whether the particles should interact only with the edges of the container.
   final bool onlyInteractWithEdges;
 
+  /// Specifies which edges of the container are hard (solid) boundaries.
+  final SolidEdges solidEdges;
+
   @override
   RelativisticEffectConfiguration copyWith({
     RelativisticEffectConfiguration Function(
       Effect<RelativisticParticle, RelativisticEffectConfiguration>,
     )? configurationOverrider,
     Gravity? gravity,
-    SolidEdges? hardEdges,
     Density? maxDensity,
     Friction? maxFriction,
     Restitution? maxRestitution,
@@ -209,21 +254,21 @@ class RelativisticEffectConfiguration extends EffectConfiguration {
     bool? foreground,
     double? maxAngle,
     double? maxBeginScale,
-    Duration? maxDuration,
     double? maxDistance,
     double? maxEndScale,
     double? maxFadeInThreshold,
     double? maxFadeOutThreshold,
     Offset? maxOriginOffset,
+    Duration? maxParticleLifespan,
     Velocity? maxVelocity,
     double? minAngle,
     double? minBeginScale,
-    Duration? minDuration,
     double? minDistance,
     double? minEndScale,
     double? minFadeInThreshold,
     double? minFadeOutThreshold,
     Offset? minOriginOffset,
+    Duration? minParticleLifespan,
     Velocity? minVelocity,
     bool? onlyInteractWithEdges,
     Offset? origin,
@@ -232,13 +277,13 @@ class RelativisticEffectConfiguration extends EffectConfiguration {
     int? particleCount,
     int? particlesPerEmit,
     Curve? scaleCurve,
+    SolidEdges? solidEdges,
     Duration? startDelay,
     Trail? trail,
   }) {
     return RelativisticEffectConfiguration(
       configurationOverrider: configurationOverrider ?? this.configurationOverrider,
       gravity: gravity ?? this.gravity,
-      hardEdges: hardEdges ?? this.hardEdges,
       maxDensity: maxDensity ?? this.maxDensity,
       maxFriction: maxFriction ?? this.maxFriction,
       maxRestitution: maxRestitution ?? this.maxRestitution,
@@ -253,18 +298,18 @@ class RelativisticEffectConfiguration extends EffectConfiguration {
       fadeOutCurve: fadeOutCurve ?? this.fadeOutCurve,
       maxAngle: maxAngle ?? this.maxAngle,
       maxBeginScale: maxBeginScale ?? this.maxBeginScale,
-      maxDuration: maxDuration ?? this.maxDuration,
       maxEndScale: maxEndScale ?? this.maxEndScale,
       maxFadeInThreshold: maxFadeInThreshold ?? this.maxFadeInThreshold,
       maxFadeOutThreshold: maxFadeOutThreshold ?? this.maxFadeOutThreshold,
       maxOriginOffset: maxOriginOffset ?? this.maxOriginOffset,
+      maxParticleLifespan: maxParticleLifespan ?? this.maxParticleLifespan,
       minAngle: minAngle ?? this.minAngle,
       minBeginScale: minBeginScale ?? this.minBeginScale,
-      minDuration: minDuration ?? this.minDuration,
       minEndScale: minEndScale ?? this.minEndScale,
       minFadeInThreshold: minFadeInThreshold ?? this.minFadeInThreshold,
       minFadeOutThreshold: minFadeOutThreshold ?? this.minFadeOutThreshold,
       minOriginOffset: minOriginOffset ?? this.minOriginOffset,
+      minParticleLifespan: minParticleLifespan ?? this.minParticleLifespan,
       minVelocity: minVelocity ?? this.minVelocity,
       onlyInteractWithEdges: onlyInteractWithEdges ?? this.onlyInteractWithEdges,
       origin: origin ?? this.origin,
@@ -272,6 +317,7 @@ class RelativisticEffectConfiguration extends EffectConfiguration {
       particleLayer: particleLayer ?? this.particleLayer,
       particlesPerEmit: particlesPerEmit ?? this.particlesPerEmit,
       scaleCurve: scaleCurve ?? this.scaleCurve,
+      solidEdges: solidEdges ?? this.solidEdges,
       startDelay: startDelay ?? this.startDelay,
       trail: trail ?? this.trail,
     );
@@ -305,7 +351,7 @@ class RelativisticEffectConfiguration extends EffectConfiguration {
           runtimeType == other.runtimeType &&
           configurationOverrider == other.configurationOverrider &&
           gravity == other.gravity &&
-          hardEdges == other.hardEdges &&
+          solidEdges == other.solidEdges &&
           maxDensity == other.maxDensity &&
           maxFriction == other.maxFriction &&
           maxRestitution == other.maxRestitution &&
@@ -321,7 +367,7 @@ class RelativisticEffectConfiguration extends EffectConfiguration {
       super.hashCode ^
       configurationOverrider.hashCode ^
       gravity.hashCode ^
-      hardEdges.hashCode ^
+      solidEdges.hashCode ^
       maxDensity.hashCode ^
       maxFriction.hashCode ^
       maxRestitution.hashCode ^